finished db auto-updater and db version 1.3, added support on response for downloading data

+++ src/main/java/edu/pitt/dbmi/daquery/update/DBVersion.java
@@ -10,7 +10,7 @@ import java.lang.annotation.Target;
 public @interface DBVersion {
 
 	//should ignore this test?
-	public int version();
+	public float version();
 	public String ddlFile();
 
 }
